requirements.txt
openai
ccxt
requests
python-dotenv
streamlit
psycopg2-binary
pandas